i/3670 Fix redundant kotlin declarations

+++ project/src/main/java/com/expedia/bookings/widget/itin/ItinCard.java
@@ -950,7 +950,7 @@ public class ItinCard<T extends ItinCardData> extends RelativeLayout
 
 			Rect headerImageVisibleRect = new Rect();
 			if (getLocalVisibleRect(headerImageVisibleRect)) {
-				percent = (float) headerImageVisibleRect.height() / (float) headerImageHeight;
+				percent = (float) headerImageVisibleRect.height() / headerImageHeight;
 			}
 
 			percent = Math.min(1.0f, Math.max(0.5f, percent));
